Clean sustainable-path: remove broken LFS files and align branch for push

+++ backend/app/models/base.py
@@ -4,7 +4,7 @@ Base model with common fields for all models
 
 from datetime import UTC, datetime
 
-from sqlalchemy import Column, DateTime, Integer
+from sqlalchemy import BigInteger, Column, DateTime
 
 from app.core.database import Base
 
@@ -14,6 +14,6 @@ class BaseModel(Base):
 
     __abstract__ = True
 
-    id = Column(Integer, primary_key=True, index=True)
+    id = Column(BigInteger, primary_key=True, index=True)
     created_at = Column(DateTime(timezone=True), default=lambda: datetime.now(UTC), nullable=False)
     updated_at = Column(DateTime(timezone=True), default=lambda: datetime.now(UTC), onupdate=lambda: datetime.now(UTC), nullable=False)
